+++ CHANGELOG.md
@@ -7,6 +7,14 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 -   add stretch option (default: true) to TextField element
 -   add multiline option (default: false) to TextField element
 
+## [2.1.0] 2016-12-01
+### Added
+-   SelectBox now has the `optionsOnTop` property (boolean, default `false`). If set to `true`, the dropdown will open to the top
+-   Pagination now has the `isTopPagination` property (boolean, default `false`). If set to `true`, the limit selector will open to the bottom
+
+### Changed
+-   SelectBox uses MDL floating label pattern
+
 ## [2.0.0] 2016-11-28
 ### Breaking
 -   Removed `<Dialog>`. Please use `<ConfirmationDialog>` or `<BaseDialog>` instead: